[pvusb] Fix cancel URB logic and unconditional notify.
A bunch of the woes the new driver was having were due to OXT-311 and
the backend not doing the ring dance right. There is also cancel logic
that needs to go into the backend driver but that can wait until the
move to 3.19 since the usbback patches are changing.

OXT-38

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>