[pvusb] Prevent module unload under normal circumstances.
Normally this driver has a module ref count of 0 on it because nothing
external references it. If it is unloaded while there are xenbus devices
running around inside of it, mayhem ensues. Now the init code takes a
ref count on the module preventing it being unloaded. The driver attr
allows you disable this if you really want to unload it (e.g. it is
convenient for development work).

OXT-38

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>