[vusb] Fix Smart card reader problems and general error handling.
The problem was seen with Smart card readers but it was really just a symptom
of a bigger bug in the front end driver. The code was not setting the length
of the returned request in the URB on unsuccessful status return codes. This
is incorrect behavior - the handling of errors should be left to the device
driver for the particular devices.

OXT-411

Signed-off-by: Ross Philipson <philipsonr@ainfosec.com>