vusb: Remove DEBUG Makefile option
DEBUG=y does not compile on Fedora 33 (v5.10.23):
./include/linux/compiler-gcc.h:102:38: error: impossible constraint in ‘asm’
  102 | #define asm_volatile_goto(x...) do { asm goto(x); asm (""); } while (0)

Passing -O0 causes the failure, and passing our own optimization is
questionable.  The kernel sets the optimization level itself, so DEBUG
or not, it should be respected.

Even regular builds generate debug info with -g, so that is redundant.

Passing -fno-omit-frame-pointer is suspect as well so drop that.

Remove DEBUG and only have a single "all" target.  This also drops -O2
from the target, but  should be provided by the kernel config anyway.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>