[usb] Fix hotplug lock-up
A race can occur during the initial communication between
the frontend and backend. On device initialization, the frontend
sends a request to the backend to get the device configuration,
namely the speed. The frontend will block until the speed of
the device struct is set to a valid value, ie anything but -1.
If the device is removed after the request is sent, but before
the backend responds, the backend will respond with -1 as the
speed (since the device is now NULL). This causes the frontend
to block indefinitely and is unrecoverable.

This commit better handles this condition in the frontend by
setting the speed to UNKNOWN if the response from the backend
is -1. Then, later in the initialization process, the frontend
will abort the operation entirely if the speed is UNKNOWN.

An important note about this change is that it now treats invalid
speed values, UNKNOWN included, as fatal. Previously, we'd issue
a warning and set the speed to HIGH. This is likely better behavior
in the long run as it doesn't blatantly ignore initialization issues.

Signed-off-by: Nicholas Tsirakis <niko.tsirakis@gmail.com>